Fixed fold and moved functions around in main read yeast / human file…

+++ asodesigner/fold.py
@@ -8,6 +8,7 @@ import subprocess
 from Bio.Seq import Seq
 from typing import Dict, List
 from enum import Enum
+from numba import njit
 
 from asodesigner.consts import RISEARCH1_BINARY_PATH
 
@@ -51,15 +52,36 @@ note: play with the d and s parameters I passed here, and with other parameters 
 '''
 
 
+def parse_mfe_scores_2(result):
+    lines = result.split('\n')
+    target_to_energies = dict()
+    for line in lines:
+        if line == '':
+            continue
+        line_parts = line.split('\t')
+        target_name = line_parts[3]
+        target_energy = line_parts[-1]
+        if line_parts[3] in target_to_energies:
+            target_to_energies[target_name].append(float(target_energy))
+        else:
+            target_to_energies[target_name] = [float(target_energy)]
+
+    # ignore the keys at this point
+    return list(target_to_energies.values())
+
+
 # in my case I was only interested in the energy scores and didn't care about the actual sequences, this is the parsing I used in case it helps you
-def get_mfe_scores(result: str) -> List[List[float]]:
+def get_mfe_scores(result: str, parsing_type=None) -> List[List[float]]:
     mfe_results = []
 
-    for gene_result in result.split("\n\nquery trigger")[1:]:
-        stripped_result = gene_result.strip()
-        regex_results = re.findall("Free energy \[kcal/mol\]: [0-9-.]+ ", stripped_result)
-        mfe_results.append(
-            [float(regex_result.replace('Free energy [kcal/mol]: ', '').strip()) for regex_result in regex_results])
+    if parsing_type is None:
+        for gene_result in result.split("\n\nquery trigger")[1:]:
+            stripped_result = gene_result.strip()
+            regex_results = re.findall("Free energy \[kcal/mol\]: [0-9-.]+ ", stripped_result)
+            mfe_results.append(
+                [float(regex_result.replace('Free energy [kcal/mol]: ', '').strip()) for regex_result in regex_results])
+    elif parsing_type == '2':
+        return parse_mfe_scores_2(result)
 
     return mfe_results
 
@@ -69,18 +91,25 @@ class Interaction(Enum):
     DNA_RNA_NO_WOBBLE = "DNA_RNA_NO_WOBBLE"
     DNA_DNA = "DNA_DNA"
 
-
-def get_trigger_mfe_scores_by_risearch(trigger: str, name_to_sequence: Dict[str, str],
-                                       interaction_type: Interaction = Interaction.DNA_RNA_NO_WOBBLE,
-                                       minimum_score: int = 1200, neighborhood: int = 0) -> str:
-    hash = random.getrandbits(64)
-
-    target_filename = f"target-{hash}.fa"
-    query_filename = f"query-{hash}.fa"
-
+def dump_target_file(target_filename: str, name_to_sequence: Dict[str, str]):
     with open(target_filename, "w") as f:
         for name, sequence in name_to_sequence.items():
             f.write(f">{name}\n{sequence}\n")
+
+
+def get_trigger_mfe_scores_by_risearch(trigger: str, name_to_sequence: Dict[str, str],
+                                       interaction_type: Interaction = Interaction.DNA_RNA_NO_WOBBLE,
+                                       minimum_score: int = 900, neighborhood: int = 0, parsing_type=None, target_file_cache=None) -> str:
+    hash = random.getrandbits(64)
+
+    if target_file_cache is None:
+        target_filename = f"target-{hash}.fa"
+        dump_target_file(target_filename, name_to_sequence)
+    else:
+        target_filename = target_file_cache
+
+    query_filename = f"query-{hash}.fa"
+
 
     with open(query_filename, "w") as f:
         f.write(f">trigger\n{Seq(trigger).reverse_complement_rna()}\n")
@@ -93,13 +122,16 @@ def get_trigger_mfe_scores_by_risearch(trigger: str, name_to_sequence: Dict[str,
         raise ValueError(f"Unsupported interaction type: {interaction_type}")
 
     try:
-        result = subprocess.check_output(
-            [RISEARCH1_BINARY_PATH, "-q", query_filename, "-t", target_filename, "-s", f"{minimum_score}", "-d", "30", "-m", m, '-n', f"{neighborhood}"],
-            universal_newlines=True,
-            text=True
-        )
+        args = [RISEARCH1_BINARY_PATH, "-q", query_filename, "-t", target_filename, "-s", f"{minimum_score}",
+                "-d", "30", "-m", m, '-n', f"{neighborhood}"]
+        if parsing_type is not None:
+            args.append(f'-p{parsing_type}')
+
+        result = subprocess.check_output(args, universal_newlines=True, text=True
+                                         )
     finally:
-        os.remove(target_filename)
+        if target_file_cache is None:
+            os.remove(target_filename)
         os.remove(query_filename)
     # logger.info(f"finished evaluating a single trigger: {trigger}")
     return result
